Modification de la méthode de dl (background)
